Refactoring
- updated readme.md
- renamed module base_src
- fixed flake8 config
