+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
     rev: 4.0.1
     hooks:
       - id: flake8
-        language_version: python3.10
+        language_version: python3.8
 
   - repo: local
     hooks:
